python/requirements-shared.txt
# auto-subtitle + ai-movie-cut 공용 의존성
faster-whisper>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=10.0.0
httpx>=0.25.0
requests>=2.28.0
